Explicitly name the default settings.py as such
