+++ server/settings/__init__.py
@@ -13,7 +13,7 @@ from os import environ
 from split_settings.tools import include, optional
 
 # Managing environment via DJANGO_ENV variable:
-environ.setdefault("DJANGO_ENV", "development")
+environ.setdefault("DJANGO_ENV", "default")
 ENV = environ["DJANGO_ENV"]
 
 base_settings = [
